+++ ReleaseNotes.md
@@ -1,4 +1,16 @@
-# Beta - v0.1.4
+# Beta - v0.2.0
 
-### Misc
-Misc css, js tweaks
+### New Feature
+- Default hours setting (in the Settings page) has been added! It can be set to any number
+
+### Home
+- Organized code
+- Added the todo list to a zustand store so it can be globally accessed or modified in other pages
+
+### Settings
+- Added settings to a zustand store so it can be globally accessed for devs
+- Added a delete data button that clears the data
+- Improved settings button feedback
+
+### Miscellaneous
+- Misc. css, js tweaks